Analytics Page modifications
