+++ lib/pages/analytics/analytics_page.dart
@@ -1,14 +1,14 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:heartless/services/enums/activity_type.dart';
 import 'package:heartless/services/enums/medical_reading_type.dart';
+import 'package:heartless/shared/provider/widget_provider.dart';
 import 'package:heartless/widgets/analytics/blood_pressure_chart.dart';
-import 'package:heartless/widgets/analytics/display_mode_selector.dart';
 import 'package:heartless/widgets/analytics/general_reading_chart.dart';
 import 'package:heartless/widgets/analytics/line_default_chart.dart';
 import 'package:heartless/widgets/analytics/radial_bar_chart.dart';
+import 'package:heartless/widgets/analytics/year_month_week_selector.dart';
 import 'package:heartless/widgets/auth/custom_two_button_toggle.dart';
+import 'package:provider/provider.dart';
 
 class AnalyticsPage extends StatefulWidget {
   final String patientId;
@@ -21,10 +21,17 @@ class AnalyticsPage extends StatefulWidget {
 class _AnalyticsPageState extends State<AnalyticsPage> {
   @override
   Widget build(BuildContext context) {
+    WidgetNotifier widgetNotifier =
+        Provider.of<WidgetNotifier>(context, listen: false);
+    widgetNotifier.setAnalyticsStartDateWithoutNotifying(
+      DateTime(DateTime.now().year, DateTime.now().month, 1),
+    );
+
     return Scaffold(
       appBar: PreferredSize(
-        preferredSize: const Size.fromHeight(100),
+        preferredSize: const Size.fromHeight(160),
         child: Column(
+          // mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
             AppBar(
               centerTitle: true,
@@ -38,63 +45,63 @@ class _AnalyticsPageState extends State<AnalyticsPage> {
               padding: const EdgeInsets.symmetric(
                 horizontal: 20,
               ),
-              child: DisplayModeSelector(
-                onDateChanged: (DateTime startDate, DateTime endDate) {
-                  log('Start Date: $startDate, End Date: $endDate');
-                  // todo: make the api call to fetch data of corresponding month and year here
-                },
-              ),
+              child: YearMonthWeekSelector(),
+            ),
+            const SizedBox(height: 20),
+            TwoButtonToggle(
+              emailPhoneToggle: false,
+              leftButtonText: 'Activites',
+              rightButtonText: 'Readings',
             ),
             // const SizedBox(height: 20),
-            // TwoButtonToggle(emailPhoneToggle: false),
           ],
-        ),
-      ),
-      bottomSheet: Padding(
-        padding: const EdgeInsets.only(bottom: 0),
-        child: TwoButtonToggle(
-          emailPhoneToggle: false,
-          leftButtonText: 'Activites',
-          rightButtonText: 'Readings',
         ),
       ),
       body: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.symmetric(vertical: 6.0, horizontal: 12),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              RadialBarChart(
-                patientId: widget.patientId,
-                date: DateTime.now(),
-              ),
-              LineDefaultChart(
-                activityType: ActivityType.exercise,
-                patientId: widget.patientId,
-                date: DateTime.now(),
-              ),
-              LineDefaultChart(
-                activityType: ActivityType.medicine,
-                patientId: widget.patientId,
-                date: DateTime.now(),
-              ),
-              LineDefaultChart(
-                activityType: ActivityType.diet,
-                patientId: widget.patientId,
-                date: DateTime.now(),
-              ),
-              BloodPressureChart(
-                patientId: widget.patientId,
-                date: DateTime.now(),
-              ),
-              GeneralReadingChart(
-                patientId: widget.patientId,
-                readingType: MedicalReadingType.weight,
-                date: DateTime.now(),
-              ),
-            ],
-          ),
+          child: Consumer<WidgetNotifier>(builder: (context, value, child) {
+            return Column(
+              mainAxisAlignment: MainAxisAlignment.start,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                RadialBarChart(
+                  patientId: widget.patientId,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                LineDefaultChart(
+                  activityType: ActivityType.exercise,
+                  patientId: widget.patientId,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                LineDefaultChart(
+                  activityType: ActivityType.medicine,
+                  patientId: widget.patientId,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                LineDefaultChart(
+                  activityType: ActivityType.diet,
+                  patientId: widget.patientId,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                BloodPressureChart(
+                  patientId: widget.patientId,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                GeneralReadingChart(
+                  patientId: widget.patientId,
+                  readingType: MedicalReadingType.weight,
+                  date: widgetNotifier.analyticsStartDate,
+                ),
+                widgetNotifier.isGraphEmpty
+                    ? Container(
+                        height: 200,
+                        color: Colors.red,
+                      )
+                    : const SizedBox(),
+              ],
+            );
+          }),
         ),
       ),
     );
