+++ lib/shared/provider/widget_provider.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:heartless/services/enums/schedule_toggle_type.dart';
+import 'package:intl/intl.dart';
 
 class WidgetNotifier with ChangeNotifier {
   // for circular progress indicator
@@ -66,5 +67,58 @@ class WidgetNotifier with ChangeNotifier {
   void removeChosenDate(DateTime date) {
     _chosenDates.remove(date);
     notifyListeners();
+  }
+
+  //* for the analytics page
+  //startDate
+  DateTime _analyticsStartDate = DateTime.now();
+  void setAnalyticsStartDate(DateTime date) {
+    _analyticsStartDate = date;
+    notifyListeners();
+  }
+
+  void setAnalyticsStartDateWithoutNotifying(DateTime date) {
+    _analyticsStartDate = date;
+  }
+
+  DateTime get analyticsStartDate => _analyticsStartDate;
+
+  //selected month
+  String _analyticsSelectedMonth =
+      DateFormat('MMM').format(DateTime.now()).toUpperCase();
+
+  void setAnalyticsSelectedMonth(String month) {
+    _analyticsSelectedMonth = month;
+    notifyListeners();
+  }
+
+  void setAnalyticsSelectedMonthWithoutNotifying(String month) {
+    _analyticsSelectedMonth = month;
+  }
+
+  String get analyticsSelectedMonth => _analyticsSelectedMonth;
+
+//selected year
+  int _analyticsSelectedYear = DateTime.now().year;
+  void setAnalyticsSelectedYear(int year) {
+    _analyticsSelectedYear = year;
+    notifyListeners();
+  }
+
+  void setAnalyticsSelectedYearWithoutNotifying(int year) {
+    _analyticsSelectedYear = year;
+  }
+
+  int get analyticsSelectedYear => _analyticsSelectedYear;
+
+  bool _isGraphEmpty = false;
+  bool get isGraphEmpty => _isGraphEmpty;
+  void setGraphEmpty(bool isEmpty) {
+    _isGraphEmpty = isEmpty;
+    notifyListeners();
+  }
+
+  void setGraphEmptyWithoutNotifying(bool isEmpty) {
+    _isGraphEmpty = isEmpty;
   }
 }
