+++ lib/widgets/analytics/radial_bar_chart.dart
@@ -8,11 +8,16 @@ import 'package:syncfusion_flutter_charts/charts.dart';
 class RadialBarChart extends StatelessWidget {
   final String patientId;
   final DateTime date;
-  const RadialBarChart(
-      {super.key, required this.patientId, required this.date});
 
+  RadialBarChart({
+    super.key,
+    required this.patientId,
+    required this.date,
+  });
+  final Map<ActivityType, List<Activity>> activitiesMap = {};
   List<RadialChartData> getDataFromSnapshot(AsyncSnapshot snapshot) {
     List<RadialChartData> newChartData = [];
+
     if (snapshot.hasData &&
         snapshot.data != null &&
         snapshot.data.docs.isNotEmpty) {
@@ -21,7 +26,6 @@ class RadialBarChart extends StatelessWidget {
         activities.add(Activity.fromMap(activity.data()));
       });
 
-      Map<ActivityType, List<Activity>> activitiesMap = {};
       activitiesMap[ActivityType.medicine] = [];
       activitiesMap[ActivityType.exercise] = [];
       activitiesMap[ActivityType.diet] = [];
@@ -58,39 +62,46 @@ class RadialBarChart extends StatelessWidget {
                 snapshot.data.docs.isNotEmpty) {
               List<RadialChartData> radialChartData =
                   getDataFromSnapshot(snapshot);
-              return SfCircularChart(
-                  tooltipBehavior: TooltipBehavior(enable: true),
-                  // title: ChartTitle(text: 'Activity Completion'),
-                  title: ChartTitle(
-                    text: 'Activity Completion',
-                    textStyle: TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                      color: Theme.of(context).shadowColor,
+              if (radialChartData.isEmpty) {
+                return Center(
+                    child: Text(
+                        'Nothing to display')); // Display a Text widget when there's no data
+              } else {
+                return SfCircularChart(
+                    tooltipBehavior: TooltipBehavior(enable: true),
+                    // title: ChartTitle(text: 'Activity Completion'),
+                    title: ChartTitle(
+                      text: 'Activity Completion',
+                      textStyle: TextStyle(
+                        fontSize: 20,
+                        fontWeight: FontWeight.bold,
+                        color: Theme.of(context).shadowColor,
+                      ),
                     ),
-                  ),
-                  legend: Legend(
-                      isVisible: true,
-                      overflowMode: LegendItemOverflowMode.wrap),
-                  series: <CircularSeries>[
-                    RadialBarSeries<RadialChartData, String>(
-                        strokeColor: Colors.white,
-                        radius: "80",
-                        innerRadius: "25",
-                        // trackBorderWidth: 0,
-                        dataSource: radialChartData,
-                        pointColorMapper: (datum, index) => datum.color,
-                        xValueMapper: (RadialChartData data, _) => data.name,
-                        yValueMapper: (RadialChartData data, _) => data.value,
-                        cornerStyle: CornerStyle.bothCurve,
-                        maximumValue: 100,
-                        dataLabelSettings: const DataLabelSettings(
-                            // Renders the data label
-                            // isVisible: true,
-                            textStyle: TextStyle(fontSize: 15)))
-                  ]);
+                    legend: Legend(
+                        isVisible: true,
+                        overflowMode: LegendItemOverflowMode.wrap),
+                    series: <CircularSeries>[
+                      RadialBarSeries<RadialChartData, String>(
+                          strokeColor: Colors.white,
+                          radius: "80",
+                          innerRadius: "25",
+                          // trackBorderWidth: 0,
+                          dataSource: radialChartData,
+                          pointColorMapper: (datum, index) => datum.color,
+                          xValueMapper: (RadialChartData data, _) => data.name,
+                          yValueMapper: (RadialChartData data, _) => data.value,
+                          cornerStyle: CornerStyle.bothCurve,
+                          maximumValue: 100,
+                          dataLabelSettings: const DataLabelSettings(
+                              // Renders the data label
+                              // isVisible: true,
+                              textStyle: TextStyle(fontSize: 15)))
+                    ]);
+              }
             } else {
-              return Container();
+              //! should be looked at again
+              return Center(child: Text('No data to display'));
             }
           }),
     );
@@ -103,3 +114,5 @@ class RadialChartData {
   final double value;
   RadialChartData(this.name, this.value, this.color);
 }
+
+// <a href="https://www.freepik.com/free-vector/tiny-programmers-working-with-system-error-computer-monitor-internet-flat-vector-illustration-programming-it-digital-technology_10173948.htm#fromView=search&page=1&position=18&uuid=3976a900-cea9-4a22-b343-e89d83520eec">Image by pch.vector on Freepik</a>
